+++ src/test/java/com/zenuml/converter/SelfMessageZenUmlTest.java
@@ -18,6 +18,7 @@ public class SelfMessageZenUmlTest extends ZenUmlTestCase {
         psiToDslNodeConverter = new PsiToDslNodeConverter();
 
     }
+
     public void test_convert_to_dsl_node_selfMessage() {
         myFixture.copyDirectoryToProject("selfMessage","");
         PsiClass selfMessageClass = myFixture.findClass("selfMessage.SelfMessage");
@@ -26,29 +27,5 @@ public class SelfMessageZenUmlTest extends ZenUmlTestCase {
         SequenceDiagram rootNode = psiToDslNodeConverter.rootNode();
         rootNode.toDsl();
         assertThat(rootNode.toDsl(), is("SelfMessage.clientMethod(){\n  SelfMessage.internalMethod();\n}"));
-    }
-
-    public void test_convert_to_dsl_node_differentClass() {
-        myFixture.copyDirectoryToProject("differentClass","");
-        PsiClass firstClass = myFixture.findClass("differentClass.FirstClass");
-        PsiMethod clientMethod = firstClass.findMethodsByName("clientMethod", true)[0];
-
-        clientMethod.accept(psiToDslNodeConverter);
-
-        SequenceDiagram rootNode = psiToDslNodeConverter.rootNode();
-        rootNode.toDsl();
-        assertThat(rootNode.toDsl(), is("FirstClass.clientMethod(){\n  SecondClass.method1();\n}"));
-    }
-
-    public void test_convert_to_dsl_node_differentClass_multiple_calls() {
-        myFixture.copyDirectoryToProject("differentClass","");
-        PsiClass firstClass = myFixture.findClass("differentClass.FirstClass");
-        PsiMethod clientMethod = firstClass.findMethodsByName("clientMethod_multiple_calls", true)[0];
-
-        clientMethod.accept(psiToDslNodeConverter);
-
-        SequenceDiagram rootNode = psiToDslNodeConverter.rootNode();
-        rootNode.toDsl();
-        assertThat(rootNode.toDsl(), is("FirstClass.clientMethod_multiple_calls(){\n  SecondClass.method1();\n  SecondClass.method1();\n}"));
     }
 }
