+++ src/test/java/com/zenuml/testFramework/fixture/ZenUmlTestCase.java
@@ -11,4 +11,7 @@ public class ZenUmlTestCase extends LightCodeInsightFixtureTestCase {
         return Paths.get("src/test/data/").toAbsolutePath().toString();
     }
 
+    public void testEmptyTest() {
+
+    }
 }
