+++ src/test/resources/file4.zenuml
@@ -1,3 +1,3 @@
 RootClass.function(a1,a2){
-  assignee=RootClass.function2(a1,a2);
+  result=RootClass.function2(a1,a2);
 }
